Create Infrastructure with DTOs
