+++ LMS.Infrastructure/Dtos/TokenDto.cs
@@ -1,4 +1,4 @@
-﻿namespace LMS.Models.Dtos;
+﻿namespace LMS.Infrastructure.Dtos;
 
 public record TokenDto(string AccessToken, string RefreshToken);
