+++ LMS.Service/LMS.Service.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\LMS.Contracts\LMS.Contracts.csproj" />
+    <ProjectReference Include="..\LMS.Infrastructure\LMS.Infrastructure.csproj" />
     <ProjectReference Include="..\LMS.Service.Contracts\LMS.Service.Contracts.csproj" />
   </ItemGroup>
